# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF656 - Service Test Management\n\n### Release\
    \ : 18.5 - December 2018\n\nService Test Management API goal is to provide the\
    \ ability to manage tests of provisioned Services. \n\n### Resource \n- ServiceTest\
    \ \n- ServiceTestSpecification\n\n### Operations\nService Test Management API\
    \ performs the following operations on the resources :\n- Retrieve an entity or\
    \ a collection of entities depending on filter criteria\n- Partial update of an\
    \ entity (including updating rules)\n- Create an entity (including default values\
    \ and creation rules)\n- Delete an entity (for administration purposes)\n- Manage\
    \ notification of events"
  version: "4.0.0"
  title: "Service Test Management"
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
- "http"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
...
  /listener/serviceTestSpecificationCreateNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceTestSpecificationCreateNotification"
      description: "Example of a client listener for receiving the notification ServiceTestSpecificationCreateNotification"
      operationId: "listenToServiceTestSpecificationCreateNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceTestSpecificationCreateNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceTestSpecificationAttributeValueChangeNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceTestSpecificationAttributeValueChangeNotification"
      description: "Example of a client listener for receiving the notification ServiceTestSpecificationAttributeValueChangeNotification"
      operationId: "listenToServiceTestSpecificationAttributeValueChangeNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceTestSpecificationAttributeValueChangeNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceTestSpecificationDeleteNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceTestSpecificationDeleteNotification"
      description: "Example of a client listener for receiving the notification ServiceTestSpecificationDeleteNotification"
      operationId: "listenToServiceTestSpecificationDeleteNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceTestSpecificationDeleteNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceTestCreateNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceTestCreateNotification"
      description: "Example of a client listener for receiving the notification ServiceTestCreateNotification"
      operationId: "listenToServiceTestCreateNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceTestCreateNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceTestAttributeValueChangeNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceTestAttributeValueChangeNotification"
      description: "Example of a client listener for receiving the notification ServiceTestAttributeValueChangeNotification"
      operationId: "listenToServiceTestAttributeValueChangeNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceTestAttributeValueChangeNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
  /listener/serviceTestDeleteNotification:
    post:
      tags:
      - "notification listeners (client side)"
      summary: "Client listener for entity ServiceTestDeleteNotification"
      description: "Example of a client listener for receiving the notification ServiceTestDeleteNotification"
      operationId: "listenToServiceTestDeleteNotification"
      parameters:
      - in: "body"
        name: "data"
        description: "The event data"
        required: true
        schema:
          $ref: "#/definitions/ServiceTestDeleteNotification"
      responses:
        201:
          description: "Notified"
          schema:
            $ref: "#/definitions/EventSubscription"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "NotificationListenersClientSide"
definitions:
...
  EventSubscription:
    type: "object"
    required:
    - "callback"
    - "id"
    properties:
      id:
        type: "string"
        description: "Id of the listener"
      callback:
        type: "string"
        description: "The callback being registered."
      query:
        type: "string"
        description: "additional data to be passed"
    description: "Sets the communication endpoint address the service instance must\
      \ use to deliver notification information"
    example:
      query: "query"
      callback: "callback"
      id: "id"
...
  ServiceTestSpecificationCreateNotification:
    type: "object"
    properties:
      eventId:
        type: "string"
        description: "The identifier of the notification"
      eventTime:
        type: "string"
        format: "date-time"
        description: "Time of the event occurence"
      eventType:
        type: "string"
        description: "The type of the notification"
      fieldPath:
        type: "string"
        description: "The path identifying the object field concerned by this notification"
      resourcePath:
        type: "string"
        description: "The path identifying the resource object concerned by this notification"
      event:
        description: "The event linked to the involved resource object"
        $ref: "#/definitions/ServiceTestSpecificationCreateEvent"
    description: "The notification data structure"
    example:
      eventId: "eventId"
      resourcePath: "resourcePath"
      eventTime: "2000-01-23T04:56:07.000+00:00"
      fieldPath: "fieldPath"
      eventType: "eventType"
      event:
        serviceTestSpecification:
          '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          testMeasureDefinition:
          - metricName: "metricName"
            unitOfMeasure: "unitOfMeasure"
            validFor:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            '@type': "@type"
            captureFrequency: "captureFrequency"
            metricHref: "metricHref"
            metricDescription: "metricDescription"
            '@baseType': "@baseType"
            valueType: "valueType"
            name: "name"
            captureMethod: "captureMethod"
            capturePeriod:
              amount: ""
              '@baseType': "@baseType"
              '@type': "@type"
              units: "units"
              '@schemaLocation': "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
            thresholdRule:
            - consequence:
              - repeatAction: true
                '@baseType': "@baseType"
                validFor:
                  startDateTime: "2000-01-23T04:56:07.000+00:00"
                  endDateTime: "2000-01-23T04:56:07.000+00:00"
                prescribeAction: "prescribeAction"
                '@type': "@type"
                name: "name"
                description: "description"
                '@schemaLocation': "http://example.com/aeiou"
              - repeatAction: true
                '@baseType': "@baseType"
                validFor:
                  startDateTime: "2000-01-23T04:56:07.000+00:00"
                  endDateTime: "2000-01-23T04:56:07.000+00:00"
                prescribeAction: "prescribeAction"
                '@type': "@type"
                name: "name"
                description: "description"
                '@schemaLocation': "http://example.com/aeiou"
              conformanceComparatorLower: "conformanceComparatorLower"
              conformanceComparatorUpper: "conformanceComparatorUpper"
              '@type': "@type"
              description: "description"
              thresholdRuleSeverity: "thresholdRuleSeverity"
              '@baseType': "@baseType"
              tolerancePeriod:
                amount: ""
                '@baseType': "@baseType"
                '@type': "@type"
                units: "units"
                '@schemaLocation': "http://example.com/aeiou"
              numberOfAllowedCrossing: 0
              conformanceTargetUpper: "conformanceTargetUpper"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              conformanceTargetLower: "conformanceTargetLower"
            - consequence:
              - repeatAction: true
                '@baseType': "@baseType"
                validFor:
                  startDateTime: "2000-01-23T04:56:07.000+00:00"
                  endDateTime: "2000-01-23T04:56:07.000+00:00"
                prescribeAction: "prescribeAction"
                '@type': "@type"
                name: "name"
                description: "description"
                '@schemaLocation': "http://example.com/aeiou"
              - repeatAction: true
                '@baseType': "@baseType"
                validFor:
                  startDateTime: "2000-01-23T04:56:07.000+00:00"
                  endDateTime: "2000-01-23T04:56:07.000+00:00"
                prescribeAction: "prescribeAction"
                '@type': "@type"
                name: "name"
                description: "description"
                '@schemaLocation': "http://example.com/aeiou"
              conformanceComparatorLower: "conformanceComparatorLower"
              conformanceComparatorUpper: "conformanceComparatorUpper"
              '@type': "@type"
              description: "description"
              thresholdRuleSeverity: "thresholdRuleSeverity"
              '@baseType': "@baseType"
              tolerancePeriod:
                amount: ""
                '@baseType': "@baseType"
                '@type': "@type"
                units: "units"
                '@schemaLocation': "http://example.com/aeiou"
              numberOfAllowedCrossing: 0
              conformanceTargetUpper: "conformanceTargetUpper"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              conformanceTargetLower: "conformanceTargetLower"
          - metricName: "metricName"
            unitOfMeasure: "unitOfMeasure"
            validFor:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            '@type': "@type"
            captureFrequency: "captureFrequency"
            metricHref: "metricHref"
            metricDescription: "metricDescription"
            '@baseType': "@baseType"
            valueType: "valueType"
            name: "name"
            captureMethod: "captureMethod"
            capturePeriod:
              amount: ""
              '@baseType': "@baseType"
              '@type': "@type"
              units: "units"
              '@schemaLocation': "http://example.com/aeiou"
            '@schemaLocation': "http://example.com/aeiou"
            thresholdRule:
            - consequence:
              - repeatAction: true
                '@baseType': "@baseType"
                validFor:
                  startDateTime: "2000-01-23T04:56:07.000+00:00"
                  endDateTime: "2000-01-23T04:56:07.000+00:00"
                prescribeAction: "prescribeAction"
                '@type': "@type"
                name: "name"
                description: "description"
                '@schemaLocation': "http://example.com/aeiou"
              - repeatAction: true
                '@baseType': "@baseType"
                validFor:
                  startDateTime: "2000-01-23T04:56:07.000+00:00"
                  endDateTime: "2000-01-23T04:56:07.000+00:00"
                prescribeAction: "prescribeAction"
                '@type': "@type"
                name: "name"
                description: "description"
                '@schemaLocation': "http://example.com/aeiou"
              conformanceComparatorLower: "conformanceComparatorLower"
              conformanceComparatorUpper: "conformanceComparatorUpper"
              '@type': "@type"
              description: "description"
              thresholdRuleSeverity: "thresholdRuleSeverity"
              '@baseType': "@baseType"
              tolerancePeriod:
                amount: ""
                '@baseType': "@baseType"
                '@type': "@type"
                units: "units"
                '@schemaLocation': "http://example.com/aeiou"
              numberOfAllowedCrossing: 0
              conformanceTargetUpper: "conformanceTargetUpper"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              conformanceTargetLower: "conformanceTargetLower"
            - consequence:
              - repeatAction: true
                '@baseType': "@baseType"
                validFor:
                  startDateTime: "2000-01-23T04:56:07.000+00:00"
                  endDateTime: "2000-01-23T04:56:07.000+00:00"
                prescribeAction: "prescribeAction"
                '@type': "@type"
                name: "name"
                description: "description"
                '@schemaLocation': "http://example.com/aeiou"
              - repeatAction: true
                '@baseType': "@baseType"
                validFor:
                  startDateTime: "2000-01-23T04:56:07.000+00:00"
                  endDateTime: "2000-01-23T04:56:07.000+00:00"
                prescribeAction: "prescribeAction"
                '@type': "@type"
                name: "name"
                description: "description"
                '@schemaLocation': "http://example.com/aeiou"
              conformanceComparatorLower: "conformanceComparatorLower"
              conformanceComparatorUpper: "conformanceComparatorUpper"
              '@type': "@type"
              description: "description"
              thresholdRuleSeverity: "thresholdRuleSeverity"
              '@baseType': "@baseType"
              tolerancePeriod:
                amount: ""
                '@baseType': "@baseType"
                '@type': "@type"
                units: "units"
                '@schemaLocation': "http://example.com/aeiou"
              numberOfAllowedCrossing: 0
              conformanceTargetUpper: "conformanceTargetUpper"
              name: "name"
              '@schemaLocation': "http://example.com/aeiou"
              conformanceTargetLower: "conformanceTargetLower"
          '@type': "@type"
          name: "name"
          description: "description"
          id: "id"
          href: "href"
          '@schemaLocation': "http://example.com/aeiou"
          version: "version"
          relatedServiceSpecification:
            targetServiceSchema:
              '@baseType': "@baseType"
              '@type': "@type"
              '@schemaLocation': "@schemaLocation"
            '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
            version: "version"
...